* Introduced xtc-generate. This is like xtc-transform. It invokes a tool without an input option and can thus be used invoke tools that do not want to have an input file (but can have other arguments).
* Extracted xtc-output from xtc-io. The xtc-output strategy
can be used like xtc-io, but doesn't read the input. It can
be used in tools that just want to generate something and do
not want to have an input.

* Changed the rename-to at the end of xtc-io (in fact now
xtc-output) to copy-to. The current FILE doesn't need to be
a temporary file and thus it is not for sure that this file
can be removed. If it is temporary, the scoped temp-files
mechanism will remove it. This will influence the
performance a little bit, but copying is implemented in
native code.

svn path=/trunk/StrategoXT/; revision=1820