* spec/opt/const-prop.str: Constant propagation for term variables replaces variables with terms to which they are matched.
* spec/opt/dead-var-elim.str: Remove term variables that are not 
used in builds. Works by replacing occurrence of dead variables in
a match by a wildcard and removing variables from scopes. A subsequent
run of optimize2 eliminates matches only consisting of a wildcard (?_)
and dead builds, i.e., builds that occur prior to another build and
will thus not be used. 

The implementation is not yet complete: choices are not taken into
account and already bound variables used in a match are discarded.

svn path=/trunk/StrategoXT/; revision=3168