* Compiler driver: Renamed sc to strc and added the -O option to regulate the amount of optimization applied by the compiler.  No optimization (-O 0) means faster compilation, but slower code.  More optimization (-O 1 to 5) means slower compilation, but faster code. The optimizations at level 3 and higher are currently experimental and don't work in all cases yet. The default level is 1. Level 2 corresponds to the optimization already performed in release 0.9.2.
* Constant propagation: Not and Rec are relevant for data-flow. Do not
propagate patterns with wildcards or As patterns. This could be
improved by filtering out As patterns and by turning wildcards into
explicit variables. A final problem turned up: when propagating a
match from a pattern with an annotation; since the annotation gets
separated from the pattern during desugaring a wrong term may be
propagated. This should be solved by integrating the treatment of
annotations in the abstract syntax, that is changing Anno(Op(f, ts1),
ts2) into Op(f, ts1, ts2). This will again have considerable impact on
the compiler.

svn path=/trunk/StrategoXT/; revision=3207