* spec/opt/stratego-laws.str: Apply simplifications using innermost. Moved most useful rules to Simplify and Optimize. Scope narrowing and cleaning is not useful anymore since these were meant for constant propagation, which is now done much better by the const-prop component.
* spec/opt/opt-lib.str: Utilities for optimization components; 
currently only apply-to-bodies/1

* spec/opt/lift-definitions.str: Don't simplify

* spec/opt/inlining.str: Inline congruences before their
expansion.

* spec/opt/define-congruences.str: simplify after defining
congruences

* spec/opt/build-match-laws.str: Define all rules directly on
sequences instead of factoring them out. This entails some
duplication since each rule should be defined for Seq(s1,s2) and
for Seq(s1,Seq(s2,s3)). However, this formulation integrates
better with innermost normalization, which is now used for
simplification.

* spec/opt/dead-var-elim.str: Don't remove bound/ubound
annotations; leave this to simplify3

* spec/opt/bound-unbound-vars.str: Avoid exponential behaviour
caused by intersecting at all definitions in a let. The effect
that this has is to undefine MarkVar for variables that are
bound in some definition, but not in another. To keep the
analysis correct, all such variables are undefined prior
to propagation in the definitions and body of a let. This
makes the analysis less optimal, but the effect is probably
small. A much better analysis ('inter-procedural') would
be needed anyway to get optimal results.

svn path=/trunk/StrategoXT/; revision=3367