syntax for labeling the current scope is now
   rules( A+x : t1 -> t2 where s )

for adding a rule and labeling the scope at the same time. This can
be reduced to

   rules( A+x :- t )

To label the current scope with x and undefine rule A for term t. Further
reduction of this syntax to

   rules( A+x )

can be used to just label the current scope with label x. Note that labels
can be arbitrary terms. Note that this also allows

   rules( A.x ) and rules( A )

Not sure what that should do. Forbid by check?

This is not yet implemented in the compiler.

svn path=/trunk/StrategoXT/; revision=5826