Further tuning the design of Stratego-Core. There are now two modes for annotations. The normal way of writing annotations is pt{t1,...,tn}, where pt is a PreTerm, a term without annotation. In the compiler we will use pt{^[t1,...,tn]} to represent a PreTerm with a list of terms as annotation. In abstract syntax, these are the 'user-level' constructs:
   AnnoList : PreTerm * List(Term) -> Term
   NoAnnoList : PreTerm -> Term

and in core we have

   Anno : PreTerm * PreTerm -> Term

The second argument must be a PreTerm here to avoid having annotations
on the list constructors of this list of annotation terms. Following
the guideline that the core syntax should be a subset of the sugared
syntax, the Anno construct (pt{^pt'}) is also visible for Stratego
programmers.

Furthermore, variables and wildcards can be both PreTerms and Terms.

The nice thing with respect to previous design of Stratego-Core is that
the Anno constructor is back and the original constructors for terms
can remain unchanged; annotation syntax is orthogonal to term syntax.

On the other hand, a term is now always constructed using two constructors,
e.g., Anno(Op(c, t*), t) or AnnoList(Op(c, t1*), t2*).


svn path=/trunk/StrategoXT/; revision=6726