Implementation of 'depends on' annotation of dynamic rules. The implementation is experimental and needs further testing. The translation scheme was tested manually on two Tiger transformations, and it should not interfere with normal dynamic rules.
The implementation is achieved by means of a desugaring
in the dynamic rule lifter. For this purpose dynamic
rule lifting has been refactored:

- lifted strategy defs are collected using a dynamic
  rule instead of returning a pair of strategy and
  lifted strategy definitions

- rules(...) sections with multiple rules are reduced
  to sequences of rules(...) with single rule definitions
  (This allows use of |[ rules( ... ) ]| syntax everywhere
  instead of DynRuleDef|[ ... ]| making spec more
  readable.)

- other desugarings are defined in 'cascading' style

A dynamic rule (A.t0 : t1 -> t2 depends on t3) is 
translated to a normal dynamic rule (A : t1 -> t2)
and series of dynamic rules 

  A : DR_KILL(x) -> (t1, t3, t0)

one for each term x in the list t3. These rules serve
to reach the declared rule from any of the dependent
terms in t3.

To kill the rule when one of the dependend terms changes
meaning the desugaring also generates the strategies
new-A and kill-A, which serve to locally shadow the
rule and kill the rule, respectively.

These strategies take the <bagof-A> DR_KILL(x) 
and for each triple (t1, t3, t0) in the result kill
the rule with the left-hand side t1, i.e., (A :- t1)
and kill the kill rules (to avoid needless killing
later on) by undefining (A :- DR_KILL(y)) for each
y in t3.

There are some subtleties involved in the scoping
of these rules, which makes things interesting. That
is, this allows the correct treatment of dynamic rules
entering a local scope shadowing some variables used
in dynamic rules. Or shorter: this solves the problem
of free variable capture in dynamic rules. 

The nice aspect of this implementation is that there
is only one dynamic rule used both for the actual
rules and the kill rules. This means that intersection
and union work directly on the kill rules as well; no
need for the programmer to be aware of the extra rules. 


svn path=/trunk/StrategoXT/; revision=7592