dr-scopes(s | names) takes a list of rule names rather than a single rule name as does dr-scope
Generalized the fork-and-merge / fix-and-merge strategies
to deal with two lists of rulenames and instantiated those
to implement the combined intersection/union operations.

The separate module dynamic-rules-utils implement the
strategies new-dynamic-rules and kill-dynamic-rules, which
allow the specification of generic data-flow strategies.
For instance, the following strategies are part of a 
generic forward propagation strategy for Tiger:

  forward-prop-vardec(before, recur, after | Rs) =
    |[ var <?x> <id> := <recur> ]|
    ; where( <new-dynamic-rules(|Rs, x)>x )
    ; after
    
  forward-prop-assign(before, recur, after | Rs) =
    |[ <?|[ x ]|> := <recur> ]|
    ; where( <kill-dynamic-rules(|Rs)>x )
    ; after

These strategies are implemented in terms of dr-new
and dr-kill, which are generated by the compiler for
specific dynamic rules. Therefore, they cannot be included
in the library (limitations of separate compilation).


svn path=/trunk/StrategoXT/; revision=7639