- xtc-register definitions included in xtc-rep - using liblib and creating separate library for xtc-lib - xtc-find and xtc-set defined in terms of nested table   operations. - all 'user-interface' operations moved to xtc.str; no longer   in xtc-lib
The XTC environment is a table with administration entries
among which a Stack entry, with a list of nested tables. xtc-find
and xtc-set operations are with respect to this stack. Pointers
to nested tables are represented by RepoFile(file) terms, which
point to a file containing an XTC repository. These files are
loaded lazily, i.e., only when the table corresponding to the
repository is queried. Loading of a repository from file into
a table is memoized using dynamic rules.


svn path=/trunk/StrategoXT/; revision=8012