Support for using pkg-config to compile Stratego programs. Currently, the stratego-runtime, stratego-lib, and the aterm library are available.
Some example:
$ pkg-config --variable=strcflags dryad java-front stratego-lib
    -I /pkg/dryad/2005-04-04-15-24/share/dryad \
    -I /pkg/dryad/2005-04-04-15-24/share/sdf/dryad \
    -I /pkg/java-front/2005-03-30-14-06/share/sdf/java-front \
    -I /pkg/java-front/2005-03-30-14-06/share/java-front \
    -I /pkg/strategoxt/2005-04-14-08-21/share/stratego-lib

$ pkg-config --cflags stratego-lib
    -I/pkg/strategoxt/2005-04-14-08-21/include \
    -I/home/martin/.nix-profile/include

$ pkg-config --libs stratego-lib
    -L/pkg/strategoxt/2005-04-14-08-21/lib/stratego-lib \
    -L/home/martin/.nix-profile/lib \
    -L/pkg/strategoxt/2005-04-14-08-21/lib/srts \
    -lstratego-lib-native-opt -lstratego-runtime-opt \
    -lstratego-runtime-choice-opt -lm -lATerm-gcc

I think this works in an acceptable way. It it useful for Makefiles as
well as users. Maybe we can even add an option to the stand-alone
compiler to specify packages to compile with. Strc can use pkg-config
to find the flags for these packages (strcflags, linker flags, and
cflags).

strc --pkgs dryad java-front xtc -i file.str

The default (and initial) value can be stratego-lib and the
stratego-runtime.

Wouldn't that be great for our users?


svn path=/trunk/StrategoXT/; revision=9680