Working on nice solution to the local installation problem. Pkg-config adds a level of indirection by letting the package determine how stuff should be build. This can also be used to solve our local installation problem.
The problem with location installations is that the location of
libraries and header files is different from an installed. Currently,
this is solved by loads of dirty hacks: there is local-baseline
directory where the packages are installed with a DESTDIR. The
libraries in the build tree are copied to this DESTDIR after that, to
allowing linking with an uninstalled library. In other words: horror.

However, pkg-config has a built-in feature for using the configuration
specified in an ${pkg}-uninstalled.pc file. Thus, we can just generate
a specific configuration for uninstalled packages.

For Libs and Cflags this is quite straightforward, but the 'xtc' tool
needs more attention. The problem is that the location of xtc is fixed
in the Makefile.xt. By introducing an xtc variable in the package
configuration,, this problem is solved. Uh ... XTC anyone?

It's not finished yet, but I'm quite confident that this is going to
work and that our build-process will be cleaner after that.


svn path=/trunk/StrategoXT/; revision=9682