STR-368: Do not generate undefine- and new- code for non-dependent dynamic rules
Currently the compiler generates undefine-R and new-R strategies for
each dynamic rule R, even if no declaration for R introduces
dependencies. The advantage is that the undefine and new strategies
can be used uniformly. However, that is only useful for declaring new
scopes, and may actually be misleading. A problem is that detecting
erroneous invocations of these strategies may not be done statically;
in the generic data-flow transformation strategies, the idea is to
parameterize the strategy with the names of a bunch of rules and call
the undefine- and new- strategies by creating the call at
run-time. Thus, only at run-time will the programmer find out that a
certain undefine- or new- rule is not defined. This is probably
acceptable.

- test77: test that depend rule generates undefine- strategy
- lift-dynamic-rules: move generation of new- and undefine- strategies
  to desugaring of dependent dynamic rules to ensure that these
  strategies are not generated otherwise. The new caching feature
  ensures that they are generated at most once.

STR-353: Detect overlapping dynamic rules 
- better error message 



svn path=/trunk/StrategoXT/; revision=11779