Various arrangements for bootstrap problems:
- declare a dummy stack_ptr in the srts. Should be removed after the
next bootstrap cycle. The problem is that the baseline compiler still
uses this stack_ptr in the generated code.

- declare SSLgetAnnotations and SSLsetAnnotations in the srts. Again,
the problem is that the baseline compiler still produces code that
invokes these functions. They should be removed after the next
bootstrap cycle.

- In the new compiler, LocalPopChoice has been replaced with
PopChoice. LocalPopChoice should be dropped after the next bootstrap
cycle.

All the cases have been annotated with @todo bootstrap problem to make
it easier to find them later.

svn path=/trunk/StrategoXT/; revision=11958