STR-382:rm-annotations does not remove annotations
The back-end compiles patterns of the form !c(t*){^[]} to just a the
construction of the term with subterms t* and constructor c and no
arguments (which is equivalent to having an empty argument list).

In the case of variables !x{^[]} is wrong however, since here the
intention is to *replace* the annotations of the term bound to x
with the empty annotation list. 


svn path=/trunk/StrategoXT/; revision=11967