STR-312: Implement failure by returning NULL (instead of setjmp/longjmp)
s2c-ng implements failure handling by letting a strategy return NULL on
failure.

This required a refactoring of the TranslateStrat rules into a recursive
function TranslateStrat(|S,F) with a success and failure label as term
arguments. A guarded choice s1 < s2 + s3 is now translated to the sequence
of statements:

    ATerm ~id:x = t;
    ~stm:<translate-strat(|Next,F')>s1
    ~stm:<translate-strat(|S,F)>s2
    ~id:F' : t = ~id:x;
    ~stm:<translate-strat(|S,F)>s3

The individual actions (match, build, call) are assumed to jump to the
S or F label on success and failure. The special label Next is used to
indicate a fall trough; simplification of the C code removes all 'goto
Next;' statements.

Some improvements as side effect:

- use concrete syntax (almost) everywhere
- terms were being built with empty annotation lists; now an annotation
  is only added when it is not the empty list


svn path=/trunk/StrategoXT/; revision=12410