STR-312: Implement failure by returning NULL (instead of setjmp/longjmp)
Rewrote native part of library to support the new failure model. The
main issue was the _get_int function, which used a call to _fail,
while returning an integer. Changed this to require that argument of
_get_int is an ATermInt, which should be checked by the caller. The
macro assert_is_int(x) can be used to check this. It will 'return
NULL' when the term x is not an ATermInt. 

Changed calls to AT_getInt to ATgetInt; all these calls are guarded by
a check that ATisInt, anyway.


svn path=/trunk/StrategoXT/; revision=12413