STR-312: Implement failure by returning NULL (instead of setjmp/longjmp)
Implementing the new failure model implies radically changing the run-time
model, which requires more care in bootstrapping.

When bootstrapping (building from baseline), build the
libstrategolib.c file using the strc in the *current*
distribution. This means that everything the compiler needs (gpp,
c-tools, etc.) needs to built before stratego-lib, in that case. The
same goes for the XTC library.


svn path=/trunk/StrategoXT/; revision=12416