STR-312: Implement failure by returning NULL (instead of setjmp/longjmp)
Term arguments of defined and primitive strategies were not checked
for NULLness. This can cause a segmentation fault when such a variable
is accessed. In practice this only (seems to) happen in primitive
calls. Fixed by including a check for each argument variable that is
*not* annotated as "bound".

Refactored generation of this check into a strategy
check-unbound-args(|F,t), which transforms its argument term (a C
statement) into an if-then-else checking each unbound variable in t
for NULLness and jumping to F if there is one.


svn path=/trunk/StrategoXT/; revision=12435