STR-312: Implement failure by returning NULL (instead of setjmp/longjmp)
The AT_getString macro has changed semantics; it no longer checks
whether its argument is an ATermString. This should be done
separately. The same holds for _get_int and _get_real. I did introduce
guards for the latter, but not for the former. This should now be
repaired in the native code of the library.


svn path=/trunk/StrategoXT/; revision=12439