Native hashtable and stream functions invoke   - hashtable_from_term or   - stream_from_term
which can now return NULL if the specified hashtable or stream is not
a valid hashtable or stream. Invokers have to take this into account,
otherwise we'll have a segmentation fault.

Added a tests for invocations of fputc and fputs on terms that are not
streams (a comparable test for hashtables was broken because of these
missing tests).


svn path=/trunk/StrategoXT/; revision=12469