STR-435: warning: SDF hack coming up
The description of a test was highlyt ambiguous, since comments are
allowed before it, which is also allowed as LAYOUT. So, I needed
context-sensitive LAYOUT: between "test" and the description only true
whitespace is allowed. This is easy to define with a kernel syntax
section.

But ... The implosion is horrible, since the whitespace will be part
of it. There should be a way to indicate that symbol as LAYOUT, but
currently there isn't it. So, I'm using the new ast attributes to
construct a term with only selected children.

This feels great :P .


svn path=/trunk/StrategoXT/; revision=13237