STR-430: Fix misc. C warnings reported by GCC
Most of these problems were actual bugs: one assignment, which should
be ==, and several _fail (return NULL) in helper functions that did
not return aterms and whose result was not checked.

I hope there are not more of these helper functions where the return
value (e.g. ATerm is ignored), since that will result in segmentation
faults instead of failure.


svn path=/trunk/StrategoXT/; revision=13272