In innermost fusion, don't reduce free constructors. As an example, when normalizing with respect to the rewrite rules
  A : Plus(Zero, x) -> x
  B : Plus(Succ(x), y) -> Succ(Plus(x, y))

the constructors Zero and Succ are 'free', that is, they do not
trigger rewrites; a term Succ(t) is in normal form if t is.

The fusion transformation turns 

 innermost(A <+ B)

into something equivalent to

 bottomup(rec r(try(
     \ Plus(Zero, x) -> x \
  <+ \ Plus(Succ(x), y) -> <r> Succ(<r>Plus(x, y)) \
 )))

That is, the variables from the left-hand sides are in normal
form so don't need to be reduced after applying a rule. 

The current addition is to remove the application of <r> to Succ(...)
since we know that this will not have an effect. The new result is

 bottomup(rec r(try(
     \ Plus(Zero, x) -> x \
  <+ \ Plus(Succ(x), y) -> Succ(<r>Plus(x, y)) \
 )))

and saves invocations of <r> (and thus pattern matches) for terms that
we know are in normal form.

This is achieved using an analysis that determines all the *non-free*
constructors, based on the matches performed by the rewrite rules. All
other constructors are considered free. A 'rule' that does not start
with a constructor match leads to abandoning this transformation,
since then we're not sure which constructors are free. This is the
achilles heel of the analysis and transformation. Incompleteness leads
to discarding reductions for terms that are not free.

To do: specialize application of <r> to non-free constructors, e.g.,
<r>Plus(x, y)



svn path=/trunk/StrategoXT/; revision=13792