Finally fixed the representation of FILE* in aterms. A FILE* is now represeted as a blob. Thanks to Eelco Dolstra, this representation is very simple: it is a blob of size 0 and the void* of the blob is the FILE*. I had a complex scheme in mind, where the pointer itself is the blob data. This would require an explicit free of the blob.
The hash of a blob is not determined based on the actual data, but on
the size and the void*. If the actual data would be hashed, then this
representation would not work, since at size 0 the data would be the
same for all blobs (thanks again to Eelco Dolstra).

It might be an option to pass the size of FILE to the blob as the size
of the data. In practice, this is useless, since you will never
construct a blob at the same address that is smaller or larger than
the FILE.

(STR-35)


svn path=/trunk/StrategoXT/; revision=13910