STR-476
Temporary workaround for the libstratego-sglr build problems on Mac OS
X. Until a new release with the dynamic sglr library is available,
I've included the source of libsglr and its dependencies, which
constructs a library called libjustsglr. libstratego-sglr has an
inter-library dependency on this library so you should not refer to
libjustsglr explicitly: it will be gone as soon as possible.

Not attractive, but the most pragmatic solution for getting the stuff
working on Mac OS X.


svn path=/strategoxt/trunk/; revision=14411