----------------------------------------------------------- Added retain-all and remove-all as aliases of filter, based on the naming convetions in the Java collections. I always have to think about filter, which feels wrong to me.
-----------------------------------------------------------
Added <zip(s | ys)> xs that works with term arguments. The strategy s
should take a term argument y of ys and is applied to an x of xs.

Of course, this is more performant than the tuple based zip, but also
it is much easier to apply if the strategy s is already a strategy
that is defined using a term argument.

Example :

  foo(|x) = ...

  main =
    <zip(foo | ys)> xs

Would be

  main =
    <zip({x, y: ?(x, y); <foo(|y)> x})> (xs, ys)

with the existing zip, which looks like serious voodoo.

(imported from Dryad)

svn path=/strategoxt/trunk/; revision=14589