STR-452: parse-unit: use libsglri
parse-unit now uses libstratego-sglr, which is a *major* performance
improvement, in particular for large languages. The parse table is now
opened only once and no intermediate processes and files are used for
parsing. The new parse-unit runs testsuites in a fraction of the time
that the previous xtc-based version needed. The performance
improvement is in particular visible in java-front and aspectj-front,
which both have rather big parse tables.

Also, fixed the support for parsing single tests.


svn path=/strategoxt/trunk/; revision=15409