STR-451: parse-unit: support matching of variable number of arguments.
Testsuites can now use Foo(...) in patterns, which is useful for
testing languages where some constructors have lots of arguments that
you don't care about.

Currently, you cannot compose ... with other patterns. So, the following examples are not yet support:

  - Foo(1, ...)
  - Foo(..., 1, ...)
  - Foo(..., 1)

This might change in the future if there is a need for this.


svn path=/strategoxt/trunk/; revision=15410