STR-119 Replace GCC nested functions by stack-allocated closures as structs
Migrating Stratego run-time to support new compilation scheme. Removing
some obsolete stuff while at it.

Main change is in stratego.h: definition of datastructure for representation
of (stack-allocated) closures and a couple of macros for initialization and
access.

Impact on functions in the runtime library (stratego.c) is minor, but affects
the signatures of the functions.

Removed the following obsolete files

  stratego-libraries/runtime/lib/stratego-choice.c
  stratego-libraries/runtime/lib/srts/stratego-choice.h

left-overs from the setjmp/longjmp era

  stratego-libraries/runtime/lib/srts/mprotect.h
  stratego-libraries/runtime/lib/mprotect.c
  stratego-libraries/config/mprotect.m4

mprotect was needed to support nested functions on platforms that
forbid executable code on the stack


svn path=/strategoxt/branches/strategoxt-sl/; revision=15470