STR-119 Replace GCC nested functions by stack-allocated closures as structs
Moving to make the static-link compilation scheme the default one. Use
'ng' components as defaults in the test directory. The compiler implementation
still uses the old components.

Cover more combinations of annotations of functions and
variables. Empty Frames with an Offset indicates a local variable that
is used in a nested function somewhere else. A call to a function with
empty Frames should be passed a pointer to the current frame.


svn path=/strategoxt/branches/strategoxt-sl/; revision=15471