STR-119 Replace GCC nested functions by stack-allocated closures as structs
Strategy arguments of primitives were not translated according to the
new scheme.

id and fail are translated to _Id(|) and _Fail(|) when used as
arguments of strategies to avoid spurious creation of identical
functions.

Higher-order strategies passed by name, cannot be lifted into a new
function. Therefore, functions created for lifting strategy arguments
are annotated as ClosureAlloc()s if their body consists of just a call
of the form f(|). Such definitions can be used to create closures
using a pointer to f instead of to the new function, thus avoiding the
creation of the function altogether. This is also useful for closures
of top-level nullary functions, which saves an intermediate function.

MkClosure has been extended to deal with plain closure
allocations. The case has to be treated separately from the regular
closure allocation case since the static link stored in the closure is
the one for the function called in the body of the closure
definition. Since this static link is taken one level higher than
assumed by escaping variables analysis, the static link is computed
from the tail of the frames list. By the nature of canonicalization,
which creates ClosureAllocs, it is to be expected that the frames list
should be at least 2 long, however, this does not appear to be the
case in practice; I'm not sure if this is an error or a
misunderstanding. Occurrences are flagged by MkClosure for
investigation.

The Stratego Library now compiles, although the code for primitives
still needs to be adapted to use StrCL as strategy arguments.


svn path=/strategoxt/branches/strategoxt-sl/; revision=15497