STR-119 Replace GCC nested functions by stack-allocated closures as structs
Native implementations of external strategies should be migrated. Done in
the case of test50, which uses bar.c.

Escaping variables analysis and canonicalization for strategy arguments
of dynamic calls.

A dynamic call is currently passed a NULL pointer as static link. This
is not a problem in the current implementation, as top-level functions
get NULL as static link anyway, and dynamically called functions should
be top-level. In the future a closure for dynamically callable functions
should be registered instead.


svn path=/strategoxt/branches/strategoxt-sl/; revision=15499