STR-492: Standard C99: report all warnings and standards compliance
Rewritten the support for calling strategies dynamically by name. The
table now stores closures, not function pointers.

The problem with function pointers was mainly that you cannot cast a
function pointer to a void pointer in ISO C99. This means that
function pointers cannot be stored in an aterm hashtable.

The new Stratego closures are pointers to structs, which can of course
be stored in an aterm blob (even if there is a function pointer in
it. Duh.).

This involved changes in the runtime and in the compiler. The compiler
now allocates memory for the closures on the heap. Maybe it would be
better to have fields for all the closures of top-level
functions. These could be used by the generated C code as well
(reduces the amount of stack space?)

Also: removed all -Wall flags from Makefiles. These will soon be set
from configure.ac (including -std=c99 -Wall etc).


svn path=/strategoxt/trunk/; revision=15518